--- a/receiver/base.go
+++ b/receiver/base.go
@@ -42,9 +42,9 @@ type Base struct {
 	concatCharacter    string
 }
 
-func NewBase(logger *zap.Logger, config tags.TagConfig) Base {
-	return Base{logger: logger, Tags: config}
-}
+// func NewBase(logger *zap.Logger, config tags.TagConfig) Base {
+// 	return Base{logger: logger, Tags: config}
+// }
 
 func sendUint64Counter(send func(metric string, value float64), metric string, value *uint64) {
 	v := atomic.LoadUint64(value)
@@ -54,6 +54,15 @@ func sendUint64Counter(send func(metric string, value float64), metric string, v
 
 func sendInt64Gauge(send func(metric string, value float64), metric string, value *int64) {
 	send(metric, float64(atomic.LoadInt64(value)))
+}
+
+func (base *Base) Init(logger *zap.Logger, config tags.TagConfig, opts ...Option) {
+	base.logger = logger
+	base.Tags = config
+
+	for _, optApply := range opts {
+		optApply(base)
+	}
 }
 
 func (base *Base) isDrop(nowTime uint32, metricTime uint32) bool {