--- a/receiver/receiver.go
+++ b/receiver/receiver.go
@@ -97,11 +97,7 @@ func New(dsn string, config tags.TagConfig, opts ...Option) (Receiver, error) {
 		return nil, err
 	}
 
-	base := NewBase(zapwriter.Logger(strings.Replace(u.Scheme, "+", "_", -1)), config)
-
-	for _, optApply := range opts {
-		optApply(&base)
-	}
+	logger := zapwriter.Logger(strings.Replace(u.Scheme, "+", "_", -1))
 
 	if u.Scheme == "tcp" {
 		addr, err := net.ResolveTCPAddr("tcp", u.Host)
@@ -110,9 +106,9 @@ func New(dsn string, config tags.TagConfig, opts ...Option) (Receiver, error) {
 		}
 
 		r := &TCP{
-			Base:      base,
 			parseChan: make(chan *Buffer),
 		}
+		r.Init(logger, config, opts...)
 
 		if err = r.Listen(addr); err != nil {
 			return nil, err
@@ -127,9 +123,9 @@ func New(dsn string, config tags.TagConfig, opts ...Option) (Receiver, error) {
 		}
 
 		r := &Pickle{
-			Base:      base,
 			parseChan: make(chan []byte),
 		}
+		r.Init(logger, config, opts...)
 
 		if err = r.Listen(addr); err != nil {
 			return nil, err
@@ -144,9 +140,9 @@ func New(dsn string, config tags.TagConfig, opts ...Option) (Receiver, error) {
 		}
 
 		r := &UDP{
-			Base:      base,
 			parseChan: make(chan *Buffer),
 		}
+		r.Init(logger, config, opts...)
 
 		if err = r.Listen(addr); err != nil {
 			return nil, err
@@ -160,9 +156,8 @@ func New(dsn string, config tags.TagConfig, opts ...Option) (Receiver, error) {
 			return nil, err
 		}
 
-		r := &GRPC{
-			Base: base,
-		}
+		r := &GRPC{}
+		r.Init(logger, config, opts...)
 
 		if err = r.Listen(addr); err != nil {
 			return nil, err
@@ -176,9 +171,8 @@ func New(dsn string, config tags.TagConfig, opts ...Option) (Receiver, error) {
 			return nil, err
 		}
 
-		r := &PrometheusRemoteWrite{
-			Base: base,
-		}
+		r := &PrometheusRemoteWrite{}
+		r.Init(logger, config, opts...)
 
 		if err = r.Listen(addr); err != nil {
 			return nil, err
@@ -192,9 +186,8 @@ func New(dsn string, config tags.TagConfig, opts ...Option) (Receiver, error) {
 			return nil, err
 		}
 
-		r := &TelegrafHttpJson{
-			Base: base,
-		}
+		r := &TelegrafHttpJson{}
+		r.Init(logger, config, opts...)
 
 		if err = r.Listen(addr); err != nil {
 			return nil, err