--- a/receiver/telegraf_http_json_test.go
+++ b/receiver/telegraf_http_json_test.go
@@ -157,9 +157,8 @@ func BenchmarkTelegrafHttpJson5(b *testing.B) {
 			}
 		}
 	}()
-	r := &TelegrafHttpJson{
-		Base: NewBase(zapwriter.Logger("telegraf"), tags.DisabledTagConfig()),
-	}
+	r := &TelegrafHttpJson{}
+	r.Init(zapwriter.Logger("telegraf"), tags.DisabledTagConfig())
 
 	r.Base.writeChan = writeChan