```go test -benchmem -run=^$ -bench ^BenchmarkTelegrafHttpJson5 github.com/lomik/carbon-clickhouse/receiver```

````
$ benchcmp json1.txt json2.txt 
benchmark                        old ns/op     new ns/op     delta
BenchmarkTelegrafHttpJson5-6     28932         13065         -54.84%

benchmark                        old allocs     new allocs     delta
BenchmarkTelegrafHttpJson5-6     157            140            -10.83%

benchmark                        old bytes     new bytes     delta
BenchmarkTelegrafHttpJson5-6     9592          10692         +11.47%
```